+++ server/app/routes.py
@@ -117,6 +117,9 @@ def has_ability(g, ability, entity):
             return False
     return True
 
+# extension functions
+post_functions = {}
+
 @app.route("/service/list", methods=["GET"])
 @filter_request("GET/service/list")
 def list_services():
@@ -532,9 +535,10 @@ def post_file(task_id, filename):
 
 @app.route("/task/log/<string:task_id>", methods=["GET"])
 @filter_request("GET/task/log")
-@task_request
 def get_log(task_id):
     content = task.get_log(redis, taskfile_dir, task_id)
+    if 'getlog' in post_functions:
+        content = post_functions['getlog'](task_id, content)
     if content is None:
         abort(flask.make_response(
             flask.jsonify(message="cannot find log for task %s" % task_id), 404))
@@ -553,6 +557,8 @@ def append_log(task_id):
 def post_log(task_id):
     content = flask.request.get_data()
     task.set_log(redis, taskfile_dir, task_id, content)
+    if 'postlog' in post_functions:
+        post_functions['postlog'](task_id, content)
     return flask.jsonify(200)
 
 @app.route("/status", methods=["GET"])